forge_platform/backend/requirements.txt
# ForgeTrace Platform Backend

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiosqlite>=0.20.0
alembic>=1.13.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
httpx>=0.27.0
click>=8.1.0

# Caching & Queue
redis>=5.0.1
celery>=5.3.4

# AWS/Cloud
boto3>=1.34.0

# ML Integration
mlflow>=2.10.0
scikit-learn>=1.3.0

# Monitoring
prometheus-client>=0.19.0
